TECH-341: Migrate PrimaryNavItem to typescript
