--- src/components/PrimaryNav/PrimaryNav.context.d.ts
@@ -1,2 +1,0 @@
-import * as React from 'react';
-
